Enable hyperlinks to namelist records
In order to hyperlink to namelist records, the commit adds hyperrefs
to all model and initialization namelist records; to distinguish between
namelist records with the same name in two cores, we use 'irec:' and 'mrec:'
prefixes. Also, to distinguish namelist options with the same name in two
cores, we use 'inl:' and 'mnl:' prefixes.